+++ include/Utilities.h
@@ -66,7 +66,7 @@ namespace Utilities
 	bool hasSpellCdPassed(const SpellbookComponent& spellbookComp, SpellIdentifier id);
 	bool hasSpellCdPassed(const Entity& entity, SpellIdentifier id);
 	bool areAxisAligned(const Entity& a, const Entity& b, int tolerance = 0);
-	float getHpPercent(const Entity& ent);
+	int getHpPercent(const Entity& ent);
 
 	//Returns sf::Text object that is adapted to be a State's header text.
 	sf::Text createStateHeaderText(const sf::Font& font, const std::string& str);
